requirements.txt
Flask[async]
requests
aiohttp
pycryptodome
cachetools
python-dotenv
protobuf
gunicorn
discord.py